+++ src/aks-preview/azext_aks_preview/_format.py
@@ -10,6 +10,42 @@ from jmespath import compile as compile_jmes, Options
 from jmespath import functions
 
 
+def aks_addon_list_available_table_format(result):
+    def parser(entry):
+        parsed = compile_jmes("""{
+                name: name,
+                description: description
+            }""")
+        return parsed.search(entry, Options(dict_cls=OrderedDict))
+    return [parser(r) for r in result]
+
+
+def aks_addon_list_table_format(result):
+    def parser(entry):
+        parsed = compile_jmes("""{
+                name: name,
+                enabled: enabled
+            }""")
+        return parsed.search(entry, Options(dict_cls=OrderedDict))
+    return [parser(r) for r in result]
+
+
+def aks_addon_show_table_format(result):
+    def parser(entry):
+        config = ""
+        for k, v in entry["config"].items():
+            config += k + "=" + v + ";"
+        entry["config"] = config
+        parsed = compile_jmes("""{
+                name: name,
+                api_key: api_key,
+                config: config,
+                identity: identity
+            }""")
+        return parsed.search(entry, Options(dict_cls=OrderedDict))
+    return parser(result)
+
+
 def aks_agentpool_show_table_format(result):
     """Format an agent pool as summary results for display with "-o table"."""
     return [_aks_agentpool_table_format(result)]