+++ src/aks-preview/azext_aks_preview/commands.py
@@ -10,6 +10,7 @@ from ._client_factory import cf_maintenance_configurations
 from ._client_factory import cf_container_services
 from ._client_factory import cf_agent_pools
 from ._format import aks_show_table_format
+from ._format import aks_addon_list_available_table_format, aks_addon_list_table_format, aks_addon_show_table_format
 from ._format import aks_agentpool_show_table_format
 from ._format import aks_agentpool_list_table_format
 from ._format import aks_versions_table_format
@@ -84,6 +85,15 @@ def load_command_table(self, _):
         g.custom_command('add', 'aks_maintenanceconfiguration_add')
         g.custom_command('update', 'aks_maintenanceconfiguration_update')
         g.custom_command('delete', 'aks_maintenanceconfiguration_delete')
+
+    # AKS addon commands
+    with self.command_group('aks addon', managed_clusters_sdk, client_factory=cf_managed_clusters) as g:
+        g.custom_command('list-available', 'aks_addon_list_available', table_transformer=aks_addon_list_available_table_format)
+        g.custom_command('list', 'aks_addon_list', table_transformer=aks_addon_list_table_format)
+        g.custom_show_command('show', 'aks_addon_show', table_transformer=aks_addon_show_table_format)
+        g.custom_command('enable', 'aks_addon_enable', supports_no_wait=True)
+        g.custom_command('disable', 'aks_addon_disable', supports_no_wait=True)
+        g.custom_command('update', 'aks_addon_update', supports_no_wait=True)
 
     # AKS agent pool commands
     with self.command_group('aks nodepool', agent_pools_sdk, client_factory=cf_agent_pools) as g: