+++ src/spring-cloud/azext_spring_cloud/_params.py
@@ -11,8 +11,9 @@ from ._validators import (validate_env, validate_cosmos_type, validate_resource_
                           validate_name, validate_app_name, validate_deployment_name, validate_log_lines,
                           validate_log_limit, validate_log_since, validate_sku, validate_jvm_options,
                           validate_vnet, validate_vnet_required_parameters, validate_node_resource_group,
-                          validate_tracing_parameters, validate_app_insights_parameters, validate_java_agent_parameters,
-                          validate_instance_count, validate_jar)
+                          validate_tracing_parameters_asc_create, validate_tracing_parameters_asc_update,
+                          validate_app_insights_parameters, validate_instance_count, validate_java_agent_parameters,
+                          validate_jar)
 from ._utils import ApiType
 
 from .vendored_sdks.appplatform.v2020_07_01.models import RuntimeVersion, TestKeyType
@@ -33,36 +34,69 @@ def load_arguments(self, _):
         c.argument('name', options_list=[
             '--name', '-n'], help='Name of Azure Spring Cloud.')
 
+    # A refactoring work item to move validators to command level to reduce the duplications.
+    # https://dev.azure.com/msazure/AzureDMSS/_workitems/edit/11002857/
     with self.argument_context('spring-cloud create') as c:
         c.argument('location', arg_type=get_location_type(self.cli_ctx), validator=validate_location)
-        c.argument('sku', type=str, validator=validate_sku, help='Name of SKU, the value is "Basic" or "Standard"')
         c.argument('reserved_cidr_range', help='Comma-separated list of IP address ranges in CIDR format. The IP ranges are reserved to host underlying Azure Spring Cloud infrastructure, which should be 3 at least /16 unused IP ranges, must not overlap with any Subnet IP ranges.', validator=validate_vnet_required_parameters)
         c.argument('vnet', help='The name or ID of an existing Virtual Network into which to deploy the Spring Cloud instance.', validator=validate_vnet_required_parameters)
         c.argument('app_subnet', help='The name or ID of an existing subnet in "vnet" into which to deploy the Spring Cloud app. Required when deploying into a Virtual Network. Smaller subnet sizes are supported, please refer: https://aka.ms/azure-spring-cloud-smaller-subnet-vnet-docs', validator=validate_vnet_required_parameters)
         c.argument('service_runtime_subnet', options_list=['--service-runtime-subnet', '--svc-subnet'], help='The name or ID of an existing subnet in "vnet" into which to deploy the Spring Cloud service runtime. Required when deploying into a Virtual Network.', validator=validate_vnet)
         c.argument('service_runtime_network_resource_group', options_list=['--service-runtime-network-resource-group', '--svc-nrg'], help='The resource group where all network resources for Azure Spring Cloud service runtime will be created in.', validator=validate_node_resource_group)
         c.argument('app_network_resource_group', options_list=['--app-network-resource-group', '--app-nrg'], help='The resource group where all network resources for apps will be created in.', validator=validate_node_resource_group)
-        c.argument('enable_java_agent', is_preview=True, arg_type=get_three_state_flag(), help="Enable java in-process agent", validator=validate_java_agent_parameters)
+        c.argument('enable_java_agent',
+                   arg_type=get_three_state_flag(),
+                   help="Java in process agent is now GA-ed and used by default when Application Insights enabled. "
+                        "This parameter is no longer needed and will be removed in future release.",
+                   validator=validate_java_agent_parameters,
+                   deprecate_info=c.deprecate(target='--enable-java-agent', hide=True))
+        c.argument('app_insights_key',
+                   help="Connection string (recommended) or Instrumentation key of the existing Application Insights.",
+                   validator=validate_tracing_parameters_asc_create)
+        c.argument('app_insights',
+                   help="Name of the existing Application Insights in the same Resource Group. "
+                        "Or Resource ID of the existing Application Insights in a different Resource Group.",
+                   validator=validate_tracing_parameters_asc_create)
+        c.argument('sampling_rate',
+                   type=float,
+                   help="Sampling Rate of application insights. Minimum is 0, maximum is 100.",
+                   validator=validate_tracing_parameters_asc_create)
+        c.argument('disable_app_insights',
+                   arg_type=get_three_state_flag(),
+                   help="Disable Application Insights, "
+                        "if not disabled and no existing Application Insights specified with "
+                        "--app-insights-key or --app-insights, "
+                        "will create a new Application Insights instance in the same resource group.",
+                   validator=validate_tracing_parameters_asc_create)
+
     with self.argument_context('spring-cloud update') as c:
-        c.argument('sku', type=str, validator=validate_sku, help='Name of SKU, the value is "Basic" or "Standard"')
+        c.argument('app_insights_key',
+                   help="Connection string (recommended) or Instrumentation key of the existing Application Insights.",
+                   validator=validate_tracing_parameters_asc_update,
+                   deprecate_info=c.deprecate(target='az spring-cloud update --app-insights-key',
+                                              redirect='az spring-cloud app-insights update --app-insights-key',
+                                              hide=True))
+        c.argument('app_insights',
+                   help="Name of the existing Application Insights in the same Resource Group. "
+                        "Or Resource ID of the existing Application Insights in a different Resource Group.",
+                   validator=validate_tracing_parameters_asc_update,
+                   deprecate_info=c.deprecate(target='az spring-cloud update --app-insights',
+                                              redirect='az spring-cloud app-insights update --app-insights',
+                                              hide=True))
+        c.argument('disable_app_insights',
+                   arg_type=get_three_state_flag(),
+                   help="Disable Application Insights, "
+                        "if not disabled and no existing Application Insights specified with "
+                        "--app-insights-key or --app-insights, "
+                        "will create a new Application Insights instance in the same resource group.",
+                   validator=validate_tracing_parameters_asc_update,
+                   deprecate_info=c.deprecate(target='az spring-cloud update --disable-app-insights',
+                                              redirect='az spring-cloud app-insights update --disable',
+                                              hide=True))
 
     for scope in ['spring-cloud create', 'spring-cloud update']:
         with self.argument_context(scope) as c:
-            c.argument('app_insights_key',
-                       help="Instrumentation key of the existing Application Insights.",
-                       validator=validate_tracing_parameters)
-            c.argument('app_insights',
-                       help="Name of the existing Application Insights in the same Resource Group. Or Resource ID of the existing Application Insights in a different Resource Group.",
-                       validator=validate_tracing_parameters)
-            c.argument('disable_distributed_tracing',
-                       arg_type=get_three_state_flag(),
-                       help="Disable distributed tracing, if not disabled and no existing Application Insights specified with --app-insights-key or --app-insights, will create a new Application Insights instance in the same resource group.",
-                       validator=validate_tracing_parameters,
-                       deprecate_info=c.deprecate(target='--disable-distributed-tracing', redirect='--disable-app-insights', hide=True))
-            c.argument('disable_app_insights',
-                       arg_type=get_three_state_flag(),
-                       help="Disable Application Insights, if not disabled and no existing Application Insights specified with --app-insights-key or --app-insights, will create a new Application Insights instance in the same resource group.",
-                       validator=validate_tracing_parameters)
+            c.argument('sku', type=str, validator=validate_sku, help='Name of SKU, the value is "Basic" or "Standard"')
             c.argument('tags', arg_type=tags_type)
 
     with self.argument_context('spring-cloud test-endpoint renew-key') as c:
@@ -268,10 +302,11 @@ def load_arguments(self, _):
 
     with self.argument_context('spring-cloud app-insights update') as c:
         c.argument('app_insights_key',
-                   help="Instrumentation key of the existing Application Insights",
+                   help="Connection string (recommended) or Instrumentation key of the existing Application Insights.",
                    validator=validate_app_insights_parameters)
         c.argument('app_insights',
-                   help="Name of the existing Application Insights in the same Resource Group. Or Resource ID of the existing Application Insights in a different Resource Group.",
+                   help="Name of the existing Application Insights in the same Resource Group. "
+                        "Or Resource ID of the existing Application Insights in a different Resource Group.",
                    validator=validate_app_insights_parameters)
         c.argument('sampling_rate',
                    type=float,